requirements.txt
addict==2.4.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
filelock==3.11.0
fonttools==4.39.3
huggingface-hub==0.13.4
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
kiwisolver==1.4.4
Markdown==3.4.3
markdown-it-py==2.2.0
matplotlib==3.7.1
mdurl==0.1.2
mmcls==0.25.0
mmcv-full==1.7.1
mmengine==0.7.0
-e git+https://github.com/open-mmlab/mmsegmentation.git@db09c540d61b5e4280dabb4920faf99563b913a4#egg=mmsegmentation
model-index==0.1.11
nibabel==5.1.0
numpy==1.24.2
opencv-python==4.7.0.72
openmim==0.3.7
ordered-set==4.1.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
prettytable==3.6.0
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
rich==13.3.3
scipy==1.10.1
six==1.16.0
tabulate==0.9.0
termcolor==2.2.0
timm==0.6.13
torch==1.13.1+cu117
torchsummary==1.5.1
torchvision==0.14.1+cu117
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
yapf==0.32.0
zipp==3.15.0
